pyproject.toml
[tool.poetry]
name = "bnn-bg"
version = "0.1.0"
description = "Code for Bayesian Neural Networks with Backgroung Knowledge graphs"
authors = ["Abdulrahman S. Omar <xabush@singularitynet.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
numpy = "^1.26.3"
jax = {version = "^0.4.23", source = "jax-gpu-src"}
jaxlib = {version = "^0.4.23", source = "jax-gpu-src"}
optax = "^0.1.7"
dm-haiku = "^0.0.11"
tensorflow-probability = "^0.23.0"
tensorflow = "^2.14.0"
tensorflow-datasets = "^4.0.0"
torch = "^2.1.2"
pandas = "^2.1.4"
scikit-learn = "^1.3.2"
SciPy = "^1.11.4"
matplotlib = "^3.8.2"
tqdm = "^4.66.1"
seaborn = "^0.13.1"
protobuf = "3.20.*"

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"


[[tool.poetry.source]]
name = "jax-gpu-src"
url = "https://storage.googleapis.com/jax-releases/jax_cuda_releases.html"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
